feat: Use babel 7 parser and visitor (#15)
BREAKING CHANGE:

the `plugins` in parser options has become an array, and we have to put `jsx` in the default value to support JSX syntax parsing. I did not add `typescript` because the fixture contains no typescript code, and `typescript` cannot co-exist with `flow`, not sure if we need to add it.